Inline TechSphere into HeroSection
